flask_app/requirements.txt
Flask==3.0.2
gunicorn==21.2.0
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
requests==2.31.0
boto3
python-dotenv